requirements.txt
flask==3.0.0
openai==1.54.3
python-dotenv==1.0.0
